requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0,<2.0

numpy>=1.24
opencv-python>=4.8

insightface>=0.7.3,<0.9
onnxruntime>=1.17.3,<2.0
# or, if using GPU (and matching CUDA/cuDNN installed):
# onnxruntime-gpu==1.17.3

mediapipe>=0.10.14,<0.12

SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pgvector>=0.2.5
pydantic>=2.6